Working on tests after previous changes.
